jump rework
working on the feel of the jump. adds more of a response to how long you hold the button for. theres a shorter jump with a quick release of the button and a higher jump with a hold. also added a property on the orbs section so the orbs still work correctly.