Bug fix to new level sequence
moved the jump pad from last level adjustment (at around 64% complete) a little to the left and up to fix an issue where player would instantly die hitting side of floor or just fly up to spikes and then die. it was working for me better but there are still times the player misses the pad but very few.